[PORT] Ports small PRs from tg: 2019-06-20 ~ 2019-07-10 (#3620)
* Fixes itching runtime

* Fixes pai runtime.

* change in operator

* Makes logs show bee reagent name

* Fixes a sprinkles runtime

* Line break change on AI minions list (#45021)

* * added newline to roundend ai minions list

* * removed extra br and space

* * robolist and extra brs removed

* Implements tgstation/tgstation#44953

Co-authored-by: Garen Crownguard <22408776+Garen7@users.noreply.github.com>

* Fixes ore bag runtime (#44938)

* Remove unneeded get_organ_target proc (#44931)

Removes the get_organ_target proc because I'm to lazy to fully
refactor laser eyes. It is embarrassing that we have a proc like this
in the code. Also makes mob code less cursed, one proc at the time.

* Really small refactor to ear and eye protection (#44832)

I really hated to look at this code every time I opened carbon defense.

* ports tgstation/tgstation#44934

Co-authored-by: nemvar <47324920+nemvar@users.noreply.github.com>

* Stasis calls end_metabolism on reagent holder

amanitin acts on delete instead of when metabolism stops

* Simplify health sensor commands (#44687)

Health sensor no longer displays a giant window with 1 button on it,
instead can change states with alt click and use in hand.

* Change cable related magic numbers to defines (#44828)

Replaced most instances where the maximum size of a cable coil was
referenced with the MAXCOIL.

* Fix sharpness-related issues (#44830)

Turned off energy weapons can no longer cut down trees, destroy wooden
walls, harvest lavaland plants and make planks.

* Even more is_sharp checks

* kill obsoleted get_location_modifier proc

* Fixes teleport armor traitor objective

* Fix span endings (#44812)

Adds a slash to instances of <span>" --> </span>".

* Leftover fixes for span endings

* [READY] Adds a chance to spawn with a horrible tie as a scientist (#44831)

The ties now spawn on a 1/250 or .4% chance

Thanks to two nice coders for grabbing my helmet and holding my hand onto the coderbus.
Thanks for mat#3542 on discord for the suggestion
cl
add: gives scientists a chance to spawn with an awesome tie
/cl

* fixes2 (#44736)

* Even more examine fixes

* Add logging for pyroclastic slimes (#44735)

Makes it easier for admins to investigate whether a slime killing
people is a pyroclastic slime or not.

* Added geo-bypass to playsound

Added --geo-bypass

* Fix runtime when clown hulks try to open doors (#44801)

Clown hulks from the admin only simple mobs cause an error when trying
to open access restricted doors, this fixes that.

* Cable Layer signal getMobhook fix (#44715)

* Slimes now 'metabolize' reagents over time (#44268)

* Update life.dm

* Update life.dm

* Update life.dm

* Fixes a slime runtime (#44689)

* Fixes a slime runtime.

* Fixes the cause, not the symptom.

* fixes russian helmets not holding vodka (#44692)

* 1

* no need for this

* e

* dsdag

* git cherry-pick betrayed me

* Examine fixes (#44688)

* blob

* do this instead

* blob examine stuff

* Adds basic antag datum logging fallback.

* reverts git magic

Co-authored-by: nemvar <haukeschuemann@outlook.de>
Co-authored-by: AnturK <AnturK@users.noreply.github.com>
Co-authored-by: kingofkosmos <monggi@gmail.com>
Co-authored-by: Garen Crownguard <22408776+Garen7@users.noreply.github.com>
Co-authored-by: nemvar <47324920+nemvar@users.noreply.github.com>
Co-authored-by: Jonathan (JJRcop) Rubenstein <jrubcop@gmail.com>
Co-authored-by: Tlaltecuhtli <33834933+Tlaltecuhtli@users.noreply.github.com>
Co-authored-by: Denton <32391752+81Denton@users.noreply.github.com>
Co-authored-by: kittymaster0 <kittymaster0@gmail.com>
Co-authored-by: vuonojenmustaturska <naksu@youzen.ext.b2.fi>
Co-authored-by: Coconutwarrior97 <40315842+Coconutwarrior97@users.noreply.github.com>
Co-authored-by: Jessica <hpone91@gmail.com>
Co-authored-by: flufflycthu1u <peter1@garritys.org>
Co-authored-by: Dennok <Deneles@yandex.ru>
Co-authored-by: skoglol <33292112+kriskog@users.noreply.github.com>