Ports PRs suggested by Archanial (#3830)
* Fixes GC for observers and new players, and tons of loose references to stuff that shouldnt be (tgstation/tgstation#55563)

I've done this on a signal because atom_hud is applied very loosely in some contexes, and the objects themselves dont have an easy way to track back into what huds they're affected by, so this seemed the best. Properly cleans up next_time_allowed, which was missing.

Update: I discovered this is also an issue for a lot of simple mobs, such as slimes and bots and such. So fixing this is huge

* Removes a source of ian harddels, keeps mcgruffs bed discription from getting overwritten at roundstart, moves the bed claiming feature to just the dogbed typepath, none of the subtypes, this applies to buckling too (tgstation/tgstation#55158)

Removes a source of ian harddels, keeps mcgruff's bed description from getting overwritten at roundstart, moves the bed claiming feature to just the dogbed typepath, blacklisting subtypes. This applies to buckling too.

This means that a dogbed can only ever belong to one dog. Fuck you.
Remake of #54892, github doesn't like force pushes, not sure why

* Some changes suggested by lummox

* Fixes an edgecase where an area can be deleted, but still be stored inside GLOB.sortedAreas, which can cause a runtime on player related Initialize(), which can cause mobs to improperly init. No I don't know why this would happen, but it happen to me. This is an ided commit (tgstation/tgstation#56954)

Fixes an edgecase where an area can be deleted, but still be stored inside GLOB.sortedAreas, which can cause a runtime on player related Initialize(), which can cause mobs to improperly init. No I don't know why this would happen, but it happen to me.

Co-authored-by: Azarak <azarak10@gmail.com>
Co-authored-by: LemonInTheDark <58055496+LemonInTheDark@users.noreply.github.com>
Co-authored-by: oranges <email@oranges.net.nz>